+++ tests/test_basic_signal/xmake.lua
@@ -21,10 +21,10 @@ target("test", function()
     end)
 
     add_files(path.join(rtl_dir, "top.sv"))
-    set_values("cfg.top", "top")
-    set_values("cfg.lua_main", "main.lua")
+    set_values("verilua.top", "top")
+    set_values("verilua.lua_main", "main.lua")
 
     if os.getenv("NO_INTERNAL_CLOCK") then
-        set_values("cfg.no_internal_clock", "1")
+        set_values("verilua.no_internal_clock", "1")
     end
 end)