+++ tests/test_comb/xmake.lua
@@ -19,12 +19,12 @@ target("test", function()
     end)
 
     add_files("./top.sv")
-    set_values("cfg.top", "top")
-    set_values("cfg.lua_main", "main.lua")
+    set_values("verilua.top", "top")
+    set_values("verilua.lua_main", "main.lua")
 
     add_values("verilator.flags", "--timing")
 
     if os.getenv("NO_INTERNAL_CLOCK") then
-        set_values("cfg.no_internal_clock", "1")
+        set_values("verilua.no_internal_clock", "1")
     end
 end)