+++ tests/test_dpic/xmake.lua
@@ -25,8 +25,8 @@ target("top", function()
     add_files("top.sv")
     add_files("dpic.cpp")
 
-    set_values("cfg.top", "top")
-    set_values("cfg.lua_main", "./main.lua")
+    set_values("verilua.top", "top")
+    set_values("verilua.lua_main", "./main.lua")
 end)
 
 target("test", function()